Test case: picks a winner, resets, and sends money
I had a challenge where the `fulfillRandomeWords` transaction in the new Promise reverted but there was no error to indicate that on the console. The transaction simply timed out. I solved it by adding a try catch block to catch any transaction errors in the promise and log it to the console.